requirements.txt
pytest==8.3.2
playwright==1.46.0
pytest-playwright==0.5.1
python-dotenv==1.0.1